alpaca_strategies/requirements.txt
# Alpaca Multi-Strategy Trading System Requirements

alpaca-py>=0.15.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
psutil>=5.9.0
